Modified Quake to run as an Android application.
+ Add an Android Java application to interface with the Android Java runtime.
+ Add a downloader activity that enables the application to download the Quake data files from
  a web server.
+ Converted to C++. (Not nescessary, just convenient.)
+ Converted from OpenGL to OpenGL ES 1.0
+ Converted from an application to a shared library that's called from Java.
+ Support Android audio, events, and networking
+ Add minimal support for running with trackball, touch screen.
+ Remap key assignments for use with the T-mobile G1 keyboard.

Features

The UI is clunky, but servicable, at least on a T-mobile G1. You can play the
shareware levels, in both single-player and multi-player deathmatches over WiFi.

On keyboardless devices you can't do much, other than admire the attract mode and quit.

Quake actually contains two source trees: QuakeWorld and WinQuake. The difference is in
how the two games implement multiplayer. QuakeWorld is client/server, while WinQuake is
peer-to-peer. QuakeWorld has only been partially converted, and is not currently compiled
or run. We concentrated on WinQuake because it was able to reliably replay the timedemo demos.
We use timedemo in WinQuake to measure OpenGL performance.

Historical Note

This change is a roll-up of numerous changes that were made in the private Google Android
perforce9 tree during the development of Android. It doesn't seem worth trying to preserve
the individual commit history, so I'm checking all the changes in at one time. This code
snapshot was taken from the Cupcake branch.

+++ quake/src/QW/client/sbar.c
@@ -934,7 +934,7 @@ void Sbar_TeamOverlay (void)
 // draw the text
 	l = scoreboardlines;
 
-	for (i=0 ; i < scoreboardteams && y <= vid.height-10 ; i++)
+	for (i=0 ; i < scoreboardteams && y <= (int) (vid.height-10) ; i++)
 	{
 		k = teamsort[i];
 		tm = teams + k;
@@ -1055,7 +1055,7 @@ void Sbar_DeathmatchOverlay (int start)
 		y += 8;
 	}
 
-	for (i=0 ; i<l && y <= vid.height-10 ; i++)
+	for (i=0 ; i<l && y <= (int) (vid.height-10) ; i++)
 	{
 		k = fragsort[i];
 		s = &cl.players[k];
@@ -1142,7 +1142,7 @@ void Sbar_DeathmatchOverlay (int start)
 		y += skip;
 	}
 
-	if (y >= vid.height-10) // we ran over the screen size, squish
+	if (y >= (int) (vid.height-10)) // we ran over the screen size, squish
 		largegame = true;
 }
 
@@ -1207,7 +1207,7 @@ void Sbar_MiniDeathmatchOverlay (void)
 
 	x = 324;
 
-	for (/* */ ; i < scoreboardlines && y < vid.height - 8 + 1; i++)
+	for (/* */ ; i < scoreboardlines && y < (int) (vid.height - 8 + 1); i++)
 	{
 		k = fragsort[i];
 		s = &cl.players[k];
@@ -1261,13 +1261,13 @@ void Sbar_MiniDeathmatchOverlay (void)
 
 	// draw seperator
 	x += 208;
-	for (y = vid.height - sb_lines; y < vid.height - 6; y += 2)
+	for (y = vid.height - sb_lines; y < (int) (vid.height - 6); y += 2)
 		Draw_Character(x, y, 14);
 
 	x += 16;
 
 	y = vid.height - sb_lines;
-	for (i=0 ; i < scoreboardteams && y <= vid.height; i++)
+	for (i=0 ; i < scoreboardteams && y <= (int) vid.height; i++)
 	{
 		k = teamsort[i];
 		tm = teams + k;
